app/pyproject.toml
[project]
name = "mep-f10-slackbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pillow>=11.1.0",
    "python-dotenv>=1.0.1",
    "schedule>=1.2.2",
    "slack-sdk>=3.34.0",
]
